Added Stand-In Player, UI Text, & Script to Track Lives, Stars, & Power
So far the script doesn't change any of the values yet, just sends them to be displayed by the UI text.
The only UI text as of current is the persistent text for the number of lives and the number of stars.
The stand-in player is just a stationary block for now.