Added Red Coins, Power Meter, Scene Reload on Death, & Game Over Message
Red coins exist in the sample scene, and will add to a red coin count (which is shown in the debug line for now) as well as the overall "coin score" when collected.
So far, the power meter is just UI text. Like in the original, it disappears when the player is at full power.
When the player dies, the entire scene reloads except for the player's number of lives, which carries over to the next reload over.
When the player goes below zero lives, a debug message plays that tells them they've hit a game over.