Put coins in the level, started moving platforms
Put the gold and red coins in the level, as faithful to the original level as possible considering what parts of the level we have done. All the moving platforms for the parts we have done have been implemented, with animators moving them. Had some issues with the first and last set of moving blocks in the level, the first don't stay where they should and both of them need at least a few to start mid-animation. We also need to change movement so that Mario moves with platforms he's standing on if he doesn't already, some research shows that parenting whatever block he's on at the time might work.