Created New (Broken) Piranha Plant Script & Prefab
The gameobjects for the Piranha Plant are all assembled and a script guiding the plant's behavior exists. Currently the script can detect if the player is close to the plant. It also attempts to rotate the plant towards the player if they are in range, but currently this isn't working quite right.
I also added a blue material for the Thwomp on the side.