IMPORTANT: Possibly Changed Unity Build Settings! Also Pihrana Things.
I was testing to see how the game would run in WebGL, and in the process changed the build platform setting to WebGL. I'm not sure if this caused any changes in the project settings, but if so, it may cause some complications when we try to merge all our branches again. Beware (maybe)!

Really, though, the main changes were to the Pihrana Plant. I allowed it to ignore the player if the player is moving slowly enough, and added a script that allows the plant to damage the player. Right now, this damage script does damage for every frame of contact, meaning it currently crashes the game and needs fixing.