Refined Whomp Code, Altered Thwomp Values & Added New Test Scene
The Whomp works more or less the same, except now it freezes when it should crush the player, in preparation for code that will allow it to actually do so.
The Thwomp now has different times for rising and falling so as to more closely match the original code.
I made a copy of the Whomp's Fortress scene that I could use to test how my enemy prefabs would behave in the actual level.