Added Rough Coin Spin Animation & RAD New Whomp Pathfinding System
NOTE: This update includes a change to the tag manager. If anyone else made changes to it, we'll have to work out conflicts when we reconvene.
Both red and yellow coin prefabs now have coin models that spin attached to them. This spin is a little rocky-looking, and could use some work.
Whomps are now intuitive and easy to place as prefabs! The way it works is you place a Whomp prefab wherever you want a Whomp to be in the level, then place two new prefabs I created: "Whomp Starting Boundary" and "Whomp Second Boundary." You attach these to the Whomp in the inspector, and then the Whomp will automatically move from it's starting position to the Starting Boundary, then to the Second Boundary, then back to the Starting Boundary, and so on. It's totally dope.