Textures, Falling Platforms, and Skybox
Added and implemented textures for all surfaces. Can still change the textures if we decide we don't like them. Created a script for falling platforms and implemented it with the falling blocks before the first wood bridge. Made a skybox out of parts of the original level's skybox, can change it if we aren't allowed to use it.