Got Whomp to Move + Attack Properly & Added Movement to Coins + Star
The Whomp now cycles in and out of it's different states as it should, patrolling, chasing the player, falling on the player and rising back up more-or-less as it should, save for a few minor tweaks it needs.
The coins and star each have their own script that rotates them, as I was having trouble getting them to rotate properly using Maya animations.